Refactoring * Abstract Post Place Handler is added
